feat: disambiguate rower attribute labels
Rename "Skill" → "Form" in all UI labels to distinguish from
"Strength" (both had an "Intermediate" value). Weight class now
displays as Lightweight/Middleweight/Heavyweight instead of
Light/Medium/Heavy. VeryStrong displays as "Very strong".

The compact stats line on lineup cards uses abbreviated labels:
Lt/Md/Hv · Nov/Int/Mst/Exp · Wk/Int/Str/V.Str · Side. The
abbreviations are unambiguous because each category has distinct
short forms.

DB enum values are unchanged — form values still submit the
original strings (Light, Medium, Intermediate, etc.).

+++ crates/server/src/templates/solve.rs
@@ -843,7 +843,7 @@ fn rower_stats_line(r: &Rower, show_attributes: bool) -> Markup {
     if show_attributes {
         html! {
             div class="text-xs text-slate-500" {
-                (r.weight_class) " · " (r.skill) " · " (r.strength) " · " (r.side)
+                (r.weight_class.short()) " · " (r.skill.short()) " · " (r.strength.short()) " · " (r.side)
             }
         }
     } else {
